backend/requirements.txt
flask>=3.0.2
flask-socketio>=5.3.6
flask-cors>=4.0.0
python-dotenv>=1.0.1
websocket-client>=1.7.0
requests>=2.31.0
watchdog>=4.0.0
google-cloud-aiplatform==1.95.1
numpy==2.2.6
sounddevice==0.5.2 